Migrate organization_stats tests
