+++ libparsec/crates/protocol/tests/organization.rs
@@ -15,100 +15,6 @@ type OrganizationBootstrapGenerator =
     Box<dyn FnOnce(&Device, Bytes, Bytes, Bytes, Bytes) -> anonymous_cmds::AnyCmdReq>;
 
 #[parsec_test]
-fn serde_organization_stats_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   cmd: "organization_stats"
-    let raw = hex!("81a3636d64b26f7267616e697a6174696f6e5f7374617473");
-
-    let req = authenticated_cmds::organization_stats::Req;
-
-    let expected = authenticated_cmds::AnyCmdReq::OrganizationStats(req);
-
-    let data = authenticated_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let authenticated_cmds::AnyCmdReq::OrganizationStats(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = authenticated_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   active_users: 1
-    //   raw_size: 8
-    //   metaraw_size: 8
-    //   realms: 1
-    //   status: "ok"
-    //   users: 1
-    //   users_per_profile_detail: [{active:1, profile:"ADMIN", revoked:0}]
-    &hex!(
-        "87ac6163746976655f757365727301a9646174615f73697a6508ad6d657461646174615f73"
-        "697a6508a67265616c6d7301a6737461747573a26f6ba5757365727301b875736572735f70"
-        "65725f70726f66696c655f64657461696c9183a770726f66696c65a541444d494ea6616374"
-        "69766501a77265766f6b656400"
-    )[..],
-    authenticated_cmds::organization_stats::Rep::Ok {
-        data_size: 8,
-        metadata_size: 8,
-        realms: 1,
-        users: 1,
-        active_users: 1,
-        users_per_profile_detail: vec![UsersPerProfileDetailItem {
-            profile: UserProfile::Admin,
-            active: 1,
-            revoked: 0,
-        }],
-    }
-)]
-#[case::not_allowed(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   reason: "foobar"
-    //   status: "not_allowed"
-    &hex!(
-        "82a6726561736f6ea6666f6f626172a6737461747573ab6e6f745f616c6c6f776564"
-    )[..],
-    authenticated_cmds::organization_stats::Rep::NotAllowed {
-        reason: Some("foobar".to_owned())
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    )[..],
-    authenticated_cmds::organization_stats::Rep::NotFound
-)]
-fn serde_organization_stats_rep(
-    #[case] raw: &[u8],
-    #[case] expected: authenticated_cmds::organization_stats::Rep,
-) {
-    let data = authenticated_cmds::organization_stats::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = authenticated_cmds::organization_stats::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_organization_config_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
